requirements.txt
bs4
pymongo
dnspython
requests
python-dotenv
